+++ src/Entries/EntryModel.php
@@ -23,7 +23,7 @@ class EntryModel extends Eloquent
     protected $casts = [
         'date' => 'datetime',
         'data' => 'json',
-        'published' => 'bool',
+        'published' => 'boolean',
     ];
 
     protected static function boot()
@@ -31,17 +31,36 @@ class EntryModel extends Eloquent
         parent::boot();
 
         static::creating(function ($entry) {
-            $entry->{$entry->getKeyName()} = (string) Str::uuid();
+            if (empty($entry->{$entry->getKeyName()})) {
+                $entry->{$entry->getKeyName()} = (string)Str::uuid();
+            }
         });
+    }
+
+    public function author()
+    {
+        return $this->belongsTo(\App\Models\User::class, 'data->author');
     }
 
     public function origin()
     {
-        return $this->belongsTo(self::class);
+        return $this->belongsTo(static::class);
+    }
+
+    public function parent()
+    {
+        return $this->belongsTo(static::class, 'data->parent');
     }
 
     public function getAttribute($key)
     {
+        // Because the import script was importing `updated_at` into the
+        // json data column, we will explicitly reference other SQL
+        // columns first to prevent errors with that bad data.
+        if (in_array($key, EntryQueryBuilder::COLUMNS)) {
+            return parent::getAttribute($key);
+        }
+
         return Arr::get($this->getAttributeValue('data'), $key, parent::getAttribute($key));
     }
 }