+++ src/Http/Controllers/EntriesController.php
@@ -3,16 +3,21 @@
 namespace Eloquenty\Http\Controllers;
 
 use Eloquenty\Entries\Entry;
-use Exception;
+use Eloquenty\Entries\EntryRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Carbon;
+use Illuminate\Validation\ValidationException;
 use Statamic\Contracts\Entries\Entry as EntryContract;
 use Statamic\CP\Breadcrumbs;
 use Statamic\Exceptions\BlueprintNotFoundException;
 use Statamic\Facades\Site;
+use Statamic\Facades\Stache;
 use Statamic\Facades\User;
 use Statamic\Http\Controllers\CP\Collections\EntriesController as StatamicEntriesController;
 use Statamic\Http\Resources\CP\Entries\Entry as EntryResource;
 use Statamic\Query\Scopes\Filters\Concerns\QueriesFilters;
+use Statamic\Support\Arr;
+use Statamic\Support\Str;
 
 class EntriesController extends StatamicEntriesController
 {
@@ -41,10 +46,14 @@ class EntriesController extends StatamicEntriesController
         $blueprint = $collection->entryBlueprint($request->blueprint);
 
         if (!$blueprint) {
-            throw new Exception(__('A valid blueprint is required.'));
+            throw new \Exception(__('A valid blueprint is required.'));
         }
 
-        $values = [];
+        if (User::current()->cant('edit-other-authors-entries', [EntryContract::class, $collection, $blueprint])) {
+            $blueprint->ensureFieldHasConfig('author', ['visibility' => 'read_only']);
+        }
+
+        $values = \Statamic\Facades\Entry::make()->collection($collection)->values()->all();
 
         // Eloquenty: Structures are disabled for eloquenty collections
         //if ($collection->hasStructure() && $request->parent) {
@@ -56,18 +65,14 @@ class EntriesController extends StatamicEntriesController
             ->addValues($values)
             ->preProcess();
 
-        $values = $fields->values()->merge([
+        $values = collect([
             'title' => null,
             'slug' => null,
             'published' => $collection->defaultPublishState(),
-        ]);
-
-        if ($collection->dated()) {
-            $values['date'] = substr(now()->toDateTimeString(), 0, 10);
-        }
+        ])->merge($fields->values());
 
         $viewData = [
-            'title' => __('Create Entry'),
+            'title' => $collection->createLabel(),
             'actions' => [
                 // Eloquenty: Use eloquenty route for store
                 'save' => cp_route('eloquenty.collections.entries.store', [$collection->handle(), $site->handle()]),
@@ -75,25 +80,27 @@ class EntriesController extends StatamicEntriesController
             'values' => $values->all(),
             'meta' => $fields->meta(),
             'collection' => $collection->handle(),
-            'collectionHasRoutes' => ! is_null($collection->route($site->handle())),
+            'collectionCreateLabel' => $collection->createLabel(),
+            'collectionHasRoutes' => !is_null($collection->route($site->handle())),
             'blueprint' => $blueprint->toPublishArray(),
             'published' => $collection->defaultPublishState(),
             'locale' => $site->handle(),
-            'localizations' => $collection->sites()->map(function ($handle) use ($collection, $site) {
-                return [
-                    'handle' => $handle,
-                    'name' => Site::get($handle)->name(),
-                    'active' => $handle === $site->handle(),
-                    'exists' => false,
-                    'published' => false,
-                    'url' => cp_route('eloquenty.collections.entries.create', [$collection->handle(), $handle]),
-                    'livePreviewUrl' => $collection->route($handle) ? cp_route('eloquenty.collections.entries.preview.create',
-                        [$collection->handle(), $handle]) : null, // Eloquenty: Use eloquenty route for preview
-                ];
-            })->all(),
+            'localizations' => $this->getAuthorizedSitesForCollection($collection)->map(function ($handle) use ($collection, $site, $blueprint) {
+                    return [
+                        'handle' => $handle,
+                        'name' => Site::get($handle)->name(),
+                        'active' => $handle === $site->handle(),
+                        'exists' => false,
+                        'published' => false,
+                    'url' => cp_route('eloquenty.collections.entries.create', [$collection->handle(), $handle, 'blueprint' => $blueprint->handle()]),
+                    'livePreviewUrl' => $collection->route($handle) ? cp_route('eloquenty.collections.entries.preview.create', [$collection->handle(), $handle]) : null, // Eloquenty: Use eloquenty route for preview
+                    ];
+            })->values()->all(),
             'revisionsEnabled' => $collection->revisionsEnabled(),
             'breadcrumbs' => $this->breadcrumbs($collection),
             'canManagePublishState' => User::current()->can('publish ' . $collection->handle() . ' entries'),
+            'previewTargets' => $collection->previewTargets()->all(),
+            'autosaveInterval' => $collection->autosaveInterval(),
         ];
 
         if ($request->wantsJson()) {
@@ -110,11 +117,26 @@ class EntriesController extends StatamicEntriesController
 
         $blueprint = $collection->entryBlueprint($request->_blueprint);
 
-        $fields = $blueprint->fields()->addValues($request->all());
+        $data = $request->all();
 
-        $fields->validate(\Statamic\Facades\Entry::createRules($collection, $site));
+        if (User::current()->cant('edit-other-authors-entries', [EntryContract::class, $collection, $blueprint])) {
+            $data['author'] = [User::current()->id()];
+        }
 
-        $values = $fields->process()->values()->except(['slug', 'date', 'blueprint']);
+        $fields = $blueprint
+            ->ensureField('published', ['type' => 'toggle'])
+            ->fields()
+            ->addValues($data);
+
+        $fields
+            ->validator()
+            ->withRules(\Statamic\Facades\Entry::createRules($collection, $site))
+            ->withReplacements([
+                'collection' => $collection->handle(),
+                'site' => $site->handle(),
+            ])->validate();
+
+        $values = $fields->process()->values()->except(['slug', 'blueprint', 'published']);
 
         // Eloquenty: Create Eloquenty Entry
         $entry = app(Entry::class)
@@ -122,32 +144,41 @@ class EntriesController extends StatamicEntriesController
             ->blueprint($request->_blueprint)
             ->locale($site->handle())
             ->published($request->get('published'))
-            ->slug($request->slug)
-            ->data($values);
+            ->slug($this->resolveSlug($request));
 
         if ($collection->dated()) {
-            $entry->date($this->formatDateForSaving($request->date));
+            $entry->date($blueprint->field('date')->fieldtype()->augment($values->pull('date')));
         }
 
+        $entry->data($values);
+
         // Eloquenty: Structures are disabled for eloquenty collections
-        //if (($structure = $collection->structure()) && ! $collection->orderable()) {
+        //if ($structure = $collection->structure()) {
         //    $tree = $structure->in($site->handle());
+        //}
+        //if ($structure && ! $collection->orderable()) {
         //    $parent = $values['parent'] ?? null;
         //    $entry->afterSave(function ($entry) use ($parent, $tree) {
+        //        if ($parent && optional($tree->find($parent))->isRoot()) {
+        //            $parent = null;
+        //        }
+        //
         //        $tree->appendTo($parent, $entry)->save();
         //    });
         //}
 
+        $this->validateUniqueUri($entry, $tree ?? null, $parent ?? null);
+
         if ($entry->revisionsEnabled()) {
-            $entry->store([
+            $saved = $entry->store([
                 'message' => $request->message,
                 'user' => User::current(),
             ]);
         } else {
-            $entry->updateLastModified(User::current())->save();
+            $saved = $entry->updateLastModified(User::current())->save();
         }
 
-        return new EntryResource($entry);
+        return (new EntryResource($entry))->additional(['saved' => $saved]);
     }
 
     public function edit(Request $request, $collection, $entry)
@@ -158,12 +189,14 @@ class EntriesController extends StatamicEntriesController
 
         $blueprint = $entry->blueprint();
 
-        if (! $blueprint) {
-            throw new BlueprintNotFoundException($entry->value('blueprint'), 'collections/'.$collection->handle());
+        if (!$blueprint) {
+            throw new BlueprintNotFoundException($entry->value('blueprint'), 'collections/' . $collection->handle());
         }
 
+        $blueprint->setParent($entry);
+
         if (User::current()->cant('edit-other-authors-entries', [EntryContract::class, $collection, $blueprint])) {
-            $blueprint->ensureFieldHasConfig('author', ['read_only' => true]);
+            $blueprint->ensureFieldHasConfig('author', ['visibility' => 'read_only']);
         }
 
         [$values, $meta] = $this->extractFromFields($entry, $blueprint);
@@ -193,12 +226,13 @@ class EntriesController extends StatamicEntriesController
             'readOnly' => User::current()->cant('edit', $entry),
             'locale' => $entry->locale(),
             'localizedFields' => $entry->data()->keys()->all(),
+            'originBehavior' => $collection->originBehavior(),
             'isRoot' => $entry->isRoot(),
             'hasOrigin' => $hasOrigin,
             'originValues' => $originValues ?? null,
             'originMeta' => $originMeta ?? null,
             'permalink' => $entry->absoluteUrl(),
-            'localizations' => $collection->sites()->map(function ($handle) use ($entry) {
+            'localizations' => $this->getAuthorizedSitesForCollection($collection)->map(function ($handle) use ($entry) {
                 $localized = $entry->in($handle);
                 $exists = $localized !== null;
 
@@ -214,12 +248,14 @@ class EntriesController extends StatamicEntriesController
                     'url' => $exists ? $localized->editUrl() : null,
                     'livePreviewUrl' => $exists ? $localized->livePreviewUrl() : null,
                 ];
-            })->all(),
+            })->values()->all(),
             'hasWorkingCopy' => $entry->hasWorkingCopy(),
             'preloadedAssets' => $this->extractAssetsFromValues($values),
             'revisionsEnabled' => $entry->revisionsEnabled(),
             'breadcrumbs' => $this->breadcrumbs($collection),
             'canManagePublishState' => User::current()->can('publish', $entry),
+            'previewTargets' => $collection->previewTargets()->all(),
+            'autosaveInterval' => $collection->autosaveInterval(),
         ];
 
         if ($request->wantsJson()) {
@@ -242,20 +278,43 @@ class EntriesController extends StatamicEntriesController
 
         $entry = $entry->fromWorkingCopy();
 
-        $fields = $entry->blueprint()->fields()->addValues($request->except('id'));
+        $blueprint = $entry->blueprint();
 
-        $fields->validate(\Statamic\Facades\Entry::updateRules($collection, $entry));
+        $data = $request->except('id');
+
+        if (User::current()->cant('edit-other-authors-entries', [EntryContract::class, $collection, $blueprint])) {
+            $data['author'] = Arr::wrap($entry->value('author'));
+        }
+
+        $fields = $entry
+            ->blueprint()
+            ->ensureField('published', ['type' => 'toggle'])
+            ->fields()
+            ->addValues($data);
+
+        $fields
+            ->validator()
+            ->withRules(\Statamic\Facades\Entry::updateRules($collection, $entry))
+            ->withReplacements([
+                'id' => $entry->id(),
+                'collection' => $collection->handle(),
+                'site' => $entry->locale(),
+            ])->validate();
 
         $values = $fields->process()->values();
+
+        // Eloquenty: Structures are disabled for eloquenty collections
+        //$parent = $values->pull('parent');
 
         if ($explicitBlueprint = $values->pull('blueprint')) {
             $entry->blueprint($explicitBlueprint);
         }
 
-        $values = $values->except(['slug', 'date']);
+        $values = $values->except(['slug', 'published']);
 
-        // Eloquenty: Structures are disabled for eloquenty collections
-        //$parent = $values->pull('parent');
+        if ($entry->collection()->dated()) {
+            $entry->date($entry->blueprint()->field('date')->fieldtype()->augment($values->pull('date')));
+        }
 
         if ($entry->hasOrigin()) {
             $entry->data($values->only($request->input('_localized')));
@@ -263,36 +322,86 @@ class EntriesController extends StatamicEntriesController
             $entry->merge($values);
         }
 
-        $entry->slug($request->slug);
-
-        if ($entry->collection()->dated()) {
-            $entry->date($this->formatDateForSaving($request->date));
-        }
+        $entry->slug($this->resolveSlug($request));
 
         // Eloquenty: Structures are disabled for eloquenty collections
-        //if ($collection->structure() && ! $collection->orderable()) {
-        //    $entry->afterSave(function ($entry) use ($parent) {
-        //        $entry->structure()
-        //            ->in($entry->locale())
+        //if ($structure = $collection->structure()) {
+        //$tree = $entry->structure()->in($entry->locale());
+        //}
+        //if ($structure && ! $collection->orderable()) {
+        //    $this->validateParent($entry, $tree, $parent);
+        //
+        //    $entry->afterSave(function ($entry) use ($parent, $tree) {
+        //        if ($parent && optional($tree->find($parent))->isRoot()) {
+        //            $parent = null;
+        //        }
+        //
+        //        $tree
         //            ->move($entry->id(), $parent)
         //            ->save();
         //    });
         //}
 
+        $this->validateUniqueUri($entry, $tree ?? null, $parent ?? null);
+
         if ($entry->revisionsEnabled() && $entry->published()) {
-            $entry
+            $saved = $entry
                 ->makeWorkingCopy()
                 ->user(User::current())
                 ->save();
+
+            // catch any changes through RevisionSaving event
+            $entry = $entry->fromWorkingCopy();
         } else {
             if (!$entry->revisionsEnabled() && User::current()->can('publish', $entry)) {
                 $entry->published($request->published);
             }
 
-            $entry->updateLastModified(User::current())->save();
+            $saved = $entry->updateLastModified(User::current())->save();
         }
 
-        return new EntryResource($entry);
+        [$values] = $this->extractFromFields($entry, $blueprint);
+
+        return (new EntryResource($entry->fresh()))->additional([
+            'saved' => $saved,
+            'data' => [
+                'values' => $values,
+            ],
+        ]);
+    }
+
+    private function resolveSlug($request)
+    {
+        return function ($entry) use ($request) {
+            if ($request->slug) {
+                return $request->slug;
+            }
+
+            if ($entry->blueprint()->hasField('slug')) {
+                return Str::slug($request->title ?? $entry->autoGeneratedTitle(), '-', $entry->site()->lang());
+            }
+
+            return null;
+        };
+    }
+
+    private function validateUniqueUri($entry, $tree, $parent)
+    {
+        //if (!$uri = $this->entryUri($entry, $tree, $parent)) {
+        //    return;
+        //}
+
+        // Eloquenty: Use Eloquenty repository, check slug and site to be unique
+        $existing = app(EntryRepository::class)->query()
+            ->where('slug', $entry->slug())
+            ->where('site', $entry->locale())
+            ->first();
+
+        if (!$existing || $existing->id() === $entry->id()) {
+            return;
+        }
+
+        throw ValidationException::withMessages(['slug' => __('statamic::validation.unique_uri')]);
     }
 
     protected function breadcrumbs($collection)