+++ src/Http/Controllers/FrontendController.php
@@ -2,12 +2,13 @@
 
 namespace Eloquenty\Http\Controllers;
 
-use Illuminate\Http\Request;
 use Eloquenty\Entries\EntryRepository;
+use Illuminate\Http\Request;
+use Statamic\Auth\Protect\Protection;
 use Statamic\Exceptions\NotFoundHttpException;
+use Statamic\Facades\Data;
 use Statamic\Facades\Site;
 use Statamic\Http\Controllers\FrontendController as StatamicFrontendController;
-use Statamic\Statamic;
 use Statamic\Support\Str;
 
 /**
@@ -23,25 +24,22 @@ class FrontendController extends StatamicFrontendController
     public function index(Request $request)
     {
         $url = Site::current()->relativePath(
-            str_finish($request->getUri(), '/')
+            Str::finish($request->getUri(), '/')
         );
 
         if ($url === '') {
             $url = '/';
         }
 
-        if (Statamic::isAmpRequest()) {
-            $url = str_after($url, '/' . config('statamic.amp.route'));
-        }
-
         if (Str::contains($url, '?')) {
             $url = substr($url, 0, strpos($url, '?'));
         }
 
-        // Eloquenty: Skip Stache
-        //if ($data = Data::findByUri($url, Site::current()->handle())) {
-        //    return $data;
-        //}
+        if ($data = Data::findByRequestUrl($request->url())) {
+            return $data;
+        }
+
+        app(Protection::class)->protect();
 
         // Eloquenty: Use Eloquenty EntryRepository to find an entry that matches current uri
         if ($entry = app(EntryRepository::class)->query()